+++ NEWS.md
@@ -8,6 +8,7 @@
 * Added `surveyFromJson()` function to create a survey from a json file made with the visual creator.
 * Themes have been implemented. See `themeFile` argument in the `SurveyModel` class.
 * Imports in __init__.py are now explicit.
+* Fixed the `monitorInput` property so it now works as expected.
 
 ## 0.3.2